requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
APScheduler==3.10.1
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.1
h11==0.14.0
idna==3.4
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.15.1
PyJWT==2.8.0
python-slugify==8.0.1
pytz==2023.3
requests==2.31.0
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
text-unidecode==1.3
traitlets==5.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
unicode==2.9
urllib3==2.0.4
uvicorn==0.23.2
wcwidth==0.2.6
